pyproject.toml
[build-system]
requires = [
	"setuptools",
        "Mathics3-Module-Base >= 9.0.0",
        "Mathics3>=9.0.0",
        "PyICU>=2.9",
]
build-backend = "setuptools.build_meta"

[project]
name = "Mathics3-Module-PyICU"
description = 'Mathics3 Hello, World! module'
dependencies = [
        "Mathics3-Module-Base >= 9.0.0",
        "Mathics3 >= 9.0.0",
        "PyICU>=2.9",
]
requires-python = ">=3.10"
readme = "README.rst"
license = "GPL-3.0-or-later"
keywords = ["Mathematica", "Wolfram", "Interpreter", "Shell", "Math", "CAS"]
maintainers = [
    {name = "Mathics Group", email = "mathics-devel@googlegroups.com"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Interpreters",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/Mathics3/Mathics3-Module-PyICU"
Downloads = "https://github.com/Mathics3/Mathics-Module-PyICU/releases"

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools]
packages = [
    "pymathics.icu",
]

[tool.setuptools.dynamic]
version = {attr = "pymathics.icu.__version__"}
